+++ packages/expo-ui/src/jetpack-compose/Switch/index.tsx
@@ -12,8 +12,20 @@ import { createViewModifierEventListener } from '../modifiers/utils';
 type SwitchElementColors = {
   checkedThumbColor?: ColorValue;
   checkedTrackColor?: ColorValue;
+  checkedBorderColor?: ColorValue;
+  checkedIconColor?: ColorValue;
   uncheckedThumbColor?: ColorValue;
   uncheckedTrackColor?: ColorValue;
+  uncheckedBorderColor?: ColorValue;
+  uncheckedIconColor?: ColorValue;
+  disabledCheckedThumbColor?: ColorValue;
+  disabledCheckedTrackColor?: ColorValue;
+  disabledCheckedBorderColor?: ColorValue;
+  disabledCheckedIconColor?: ColorValue;
+  disabledUncheckedThumbColor?: ColorValue;
+  disabledUncheckedTrackColor?: ColorValue;
+  disabledUncheckedBorderColor?: ColorValue;
+  disabledUncheckedIconColor?: ColorValue;
 };
 
 // @docsMissing
@@ -46,6 +58,12 @@ export type SwitchProps = {
    * @default 'switch'
    */
   variant?: 'checkbox' | 'switch' | 'button';
+  /**
+   * Whether the switch is enabled.
+   * @default true
+   * @platform android
+   */
+  enabled?: boolean;
   /**
    * Callback function that is called when the checked state changes.
    */